+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changelog
 ===================
 
+## 2.0.1 - 2019-05-20
+
+### Added
+- Added a warning for Twitter redirect URIs containing unsupported query parameters.
+
+### Changed
+- Social now requires Craft CMS 3.1+. ([#31](https://github.com/dukt/social/issues/31))
+
+### Fixed
+- Fixed a bug where Social was not handling soft deletes properly, preventing users from registering with the same social credentials as a soft-deleted user. ([#32](https://github.com/dukt/social/issues/32), [#35](https://github.com/dukt/social/issues/35))
+
 ## 2.0.0 - 2019-03-04
 
 ### Added